Changed response body, string to response entity

+++ src/main/java/org/unstoppable/projectstack/controller/CommunityRestController.java
@@ -1,9 +1,13 @@
 package org.unstoppable.projectstack.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.util.UriComponentsBuilder;
 import org.unstoppable.projectstack.entity.Channel;
 import org.unstoppable.projectstack.entity.Community;
 import org.unstoppable.projectstack.entity.Subscription;
@@ -17,6 +21,7 @@ import org.unstoppable.projectstack.service.UserService;
 import org.unstoppable.projectstack.validator.CommunityValidator;
 
 import javax.validation.Valid;
+import java.net.URI;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
@@ -43,22 +48,28 @@ public class CommunityRestController {
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public String addCommunity(@Valid @RequestBody CommunityCreationForm communityForm,
-                               BindingResult result,
-                               Principal principal) {
+    public ResponseEntity<Void> addCommunity(@Valid @RequestBody CommunityCreationForm communityForm,
+                                             BindingResult result,
+                                             Principal principal,
+                                             UriComponentsBuilder uriComponentsBuilder) {
         new CommunityValidator(communityService).validate(communityForm, result);
         if (result.hasErrors()) {
-            return "failure";
+            return new ResponseEntity<>(HttpStatus.CONFLICT);
         } else {
             Community community = communityForm.createCommunity();
             communityService.save(community);
             // After community creation we should add default channel
             Channel defaultChannel = createDefaultChannel(community);
             channelService.add(defaultChannel);
-            // And subscribe creator to that community
+            // Subscribe creator to that community
             User user = userService.getByUsername(principal.getName());
             subscriptionService.subscribe(createSubscription(community, user));
-            return "success";
+            // Create header
+            HttpHeaders headers = new HttpHeaders();
+            // And header location
+            URI location = uriComponentsBuilder.path("/{communityTitle}").buildAndExpand(community.getTitle()).toUri();
+            headers.setLocation(location);
+            return new ResponseEntity<>(headers, HttpStatus.CREATED);
         }
     }
 
@@ -90,26 +101,30 @@ public class CommunityRestController {
     }
 
     @RequestMapping(value = "/join", method = RequestMethod.POST)
-    public String subscribe(String communityTitle, Principal principal) {
+    public ResponseEntity<Void> subscribe(String communityTitle, Principal principal) {
         if (principal != null) {
             Community community = communityService.getByTitle(communityTitle);
             User user = userService.getByUsername(principal.getName());
             subscriptionService.subscribe(createSubscription(community, user));
-            return "success";
+            return new ResponseEntity<>(HttpStatus.OK);
         }
-        return "failure";
+        return new ResponseEntity<>(HttpStatus.METHOD_NOT_ALLOWED);
     }
 
     @RequestMapping(value = "/leave", method = RequestMethod.POST)
-    public String unsubscribe(String communityTitle, Principal principal) {
+    public ResponseEntity<Void> unsubscribe(String communityTitle, Principal principal) {
         if (principal != null) {
             Community community = communityService.getByTitle(communityTitle);
             User user = userService.getByUsername(principal.getName());
             Subscription subscription = subscriptionService.get(community, user);
-            subscriptionService.delete(subscription);
-            return "success";
+            if (subscription != null) {
+                subscriptionService.delete(subscription);
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
-        return "failure";
+        return new ResponseEntity<>(HttpStatus.METHOD_NOT_ALLOWED);
     }
 
     private CommunitySubscription createCommunitySubscription(Community community) {
